requirements.txt
aiogram
supabase
asyncpg
python-dotenv
openai
httpx
apscheduler
openpyxl
yookassa 